// src/client/features/KbmDiscountModal/lib/driversKbmStatuses.ts
export const enum IStatuses {
  ThereAreDriversWithDefaultKbm = 'ThereAreDriversWithDefaultKbm',
  ThereAreDriversWithAccidents = 'ThereAreDriversWithAccidents',
  ThereAreDriversWithMaxKbm = 'ThereAreDriversWithMaxKbm',
  TotalMinKbm = 'TotalMinKbm',
  DefaultStatus = 'DefaultStatus',
  TheOnlyDriverLeftOrNoInsurer = 'TheOnlyDriverLeftOrNoInsurer',
  LoadingStatus = 'LoadingStatus',
}